devprod: reorganize run code
